Feat: welcome screen, session-list toolbar, test suite, and missing-dir handling
- Add empty-state welcome screen with recent projects, agent switch, and repo link
- Add session-list toolbar — search / sort / filter-by-ID with keyword highlighting
- Add "new session in terminal", refresh, and delete-project actions to the session-list header
- Add keyword-highlighted search, click-to-open detail, and hover spotlight to the Trash list
- Add fly animations for single-session restore and project deletion; extract a shared fly module
- Hide directory-dependent actions and show a "Directory missing" tag for orphaned projects
- Move toast notifications from bottom to top-center
- Toggle project selection — click the active project to return to the welcome screen
- Fix queued user messages (including image-bearing ones) being dropped from Claude transcripts
- Add a Vitest test suite and a GitHub Actions CI workflow
- Update CHANGELOG, README roadmap, and CLAUDE.md

+++ src-tauri/src/util.rs
@@ -78,7 +78,7 @@ pub fn format_iso8601_utc(secs: i64, ms: u32) -> String {
     let s = secs.rem_euclid(60) as u32;
     let m = (secs.div_euclid(60)).rem_euclid(60) as u32;
     let h = (secs.div_euclid(3600)).rem_euclid(24) as u32;
-    let mut days = secs.div_euclid(86400) as i64;
+    let mut days = secs.div_euclid(86400);
     let mut year: i64 = 1970;
     loop {
         let leap = (year % 4 == 0 && year % 100 != 0) || year % 400 == 0;
